requirements-local-3.8.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    ./bin/build.sh gen_requirements 3.8
#
amalgam-lang==2.3.18
    # via howso-engine-api
certifi==2023.7.22
    # via
    #   howso-engine
    #   howso-engine-api
cffi==1.15.1
    # via cryptography
cloudpickle==2.2.1
    # via gymnasium
cryptography==41.0.3
    # via howso-engine-api
deprecation==2.1.0
    # via howso-engine-api
exceptiongroup==1.1.3
    # via pytest
execnet==2.0.2
    # via pytest-xdist
faker==19.3.1
    # via howso-engine-api
farama-notifications==0.0.4
    # via gymnasium
flake8==3.9.2
    # via -r requirements-dev.in
gymnasium[classic-control]==0.29.1
    # via -r requirements.in
howso-engine==1.1.46
    # via
    #   -r requirements-local.in
    #   -r requirements.in
howso-engine-api==7.1.0
    # via howso-engine
howso-openapi-client==24.0.15
    # via howso-engine-api
humanize==4.8.0
    # via howso-engine-api
importlib-metadata==6.8.0
    # via gymnasium
iniconfig==2.0.0
    # via pytest
joblib==1.3.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
mccabe==0.6.1
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mmh3==4.0.1
    # via howso-engine-api
numpy==1.23.5
    # via
    #   -r requirements.in
    #   gymnasium
    #   howso-engine-api
    #   pandas
    #   scikit-learn
    #   scipy
packaging==23.1
    # via
    #   deprecation
    #   pytest
pandas==1.5.0
    # via
    #   -r requirements.in
    #   howso-engine-api
pluggy==1.3.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.3.1
    # via flake8
pygame==2.5.1
    # via gymnasium
pygments==2.16.1
    # via rich
pyjwt==2.8.0
    # via howso-engine-api
pytest==7.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-xdist
pytest-xdist==3.3.1
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   faker
    #   howso-openapi-client
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via howso-engine-api
rich==13.5.2
    # via howso-engine-api
scikit-learn==1.3.0
    # via howso-engine-api
scipy==1.10.1
    # via scikit-learn
semantic-version==2.10.0
    # via howso-engine-api
six==1.16.0
    # via
    #   howso-openapi-client
    #   python-dateutil
threadpoolctl==3.2.0
    # via scikit-learn
tomli==2.0.1
    # via pytest
typing-extensions==4.7.1
    # via
    #   faker
    #   gymnasium
    #   howso-engine-api
    #   rich
urllib3==1.26.16
    # via
    #   howso-engine-api
    #   howso-openapi-client
zipp==3.16.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==68.1.2
    # via howso-openapi-client
